requirements.txt
Django>=5.1.0,<6.0.0
djangorestframework>=3.14.0,<4.0.0
django-filter>=23.5,<24.0
django-crispy-forms>=2.0,<3.0
crispy-bootstrap5>=2023.10,<2024.0
Pillow>=10.0.0,<11.0.0
python-dotenv>=1.0.0,<2.0.0
gunicorn>=21.2.0,<22.0.0
whitenoise>=6.5.0,<7.0.0
psycopg2-binary>=2.9.9,<3.0.0  # Para PostgreSQL (opcional) 